Fixed All the Errors?
If I didn't fix all of my typos and errors (like forgetting to use $get[] and just flat out typing the name of the variable simpliciter. If I didn't fix all of the errors, then I fixed a LOT of them x). I'm excited to test out this newly finished code (for lack of a better term).